+++ store/reducers/auth.ts
@@ -1,9 +1,11 @@
 import { createSlice } from "@reduxjs/toolkit";
+import { UserAuth } from "@/types";
 
 interface initialStateType {
     phoneNumber: string;
     otp: string;
     access_token: string;
+    userInfor?: UserAuth;
 }
 
 const initialState: initialStateType = {
@@ -16,6 +18,9 @@ const authSlice = createSlice({
     name: "sign",
     initialState,
     reducers: {
+        setUserInfor: (state, { payload }) => {
+            state.userInfor = payload;
+        },
         setInfoSign: (state, { payload }) => {
             state.otp = payload.otp;
             state.phoneNumber = payload.phoneNumber;
@@ -26,6 +31,6 @@ const authSlice = createSlice({
     },
 });
 
-export const { setInfoSign, setAccessToken } = authSlice.actions;
+export const { setUserInfor, setInfoSign, setAccessToken } = authSlice.actions;
 
 export default authSlice.reducer;
